Create JSONLoader Tests and change PatternsViewModel as a exapmle for further changes
